Code Update
I made a General Knowledge Quiz Game using C programming language.

This Quiz Game shows the questions to the user along with a set of options for their answer.

The user can type in the letter of their answer and the program will show whether or no that answer is correct.

If the answer is correct, the a point of 1 will be added to their score.

If the answer is wrong, there will be no point added to their score and the program will tell them that their answer was wrong.

Then, once they completed the Quiz Game, their final score will be shown.