Quick Bug Fix
I fixed an error where i misspelled the "sQuestions" in one of the printf functions used to print the questions.

Then, I also fixed an error where I overlooked a line of code (the scanf function used to input the answer of the user), where I wasn't able to terminate the line of code.