Score and Options Bug Fix
I resolved a problem with the final score's showing of negative digits. This was resolved by setting my score variable to 0.

In an effort to stop my options from repeating, I modified the statements inside my inner for loop. But it later became clear that wasn't the problem.

I also changed the way to clear the \n within my input buffer, but then that also did not fix the issue about my answer options.

Finally, I changed the sOptions index from i to j in the inner for loop where i print the options for the question. Fortunately, that resolved the problem with my options.